refactor: extract mcp handshake logic to new file
